pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-tampering-agent"
version = "0.1.0"
description = "Composable LangGraph agent for video tampering detection"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.5",
    "langgraph>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.7",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.1.0",
    "python-docx>=1.1.2",
    "docx2pdf>=0.1.8"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "respx>=0.21"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"video_tampering_agent" = ["py.typed"]
